On en est à extraire le body des POST ds reqst_handler::reader l.75 a choisir selon les 3 possibilités marquées ds les commentaires à cet endroit : 1ere SOLUTION : mise ds la map[BODY] 2eme SOLUTION : C'EST ICI QU'ON DOIT UTILISER extract_post_rqst_body 3eme SOLUTION : SI C'EST UNE REQUESTE POST ON IGNORE LE BODY POUR USAGE ULTÉRIEUR AV _C_INFO_PTR

+++ ATwebserv/server.hpp
@@ -27,7 +27,7 @@ public:
     void run(void); // lance les étapes de configuration des sockets initialize(), de run
 
     /* PRIVATE */
-    int is_new_client(int fd);
+    size_t is_new_client(int fd);
     int get_time_out(int id_serv);
 };
